2006-03-30  Richard Guenther  <rguenther@suse.de>
	* i386/Makefile.am: Do not build dbl-64 variants.
	* i386/Makefile.in: Regenerate.
	* dbl-64: Remove.

	* config/i386/i386.c: Remove builtins for SSE2 ABI intrinsic
	variants with double arguments.

	* gcc.target/i386/sselibm-1.c: Adjust for libgcc-math partial
	removal.
	* gcc.target/i386/sselibm-2.c: Likewise.
	* gcc.target/i386/sselibm-3.c: Likewise.
	* gcc.target/i386/sselibm-4.c: Likewise.
	* gcc.target/i386/sselibm-5.c: Likewise.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@112530 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.target/i386/sselibm-2.c
@@ -2,11 +2,11 @@
 /* { dg-options "-msse2 -mfpmath=sse -msselibm" } */
 /* { dg-require-effective-target ilp32 } */
 
-double sin(double);
+float sinf(float);
 
-double foo(double x)
+float foo(float x)
 {
-  return sin(x);
+  return sinf(x);
 }
 
-/* { dg-final { scan-assembler "__libm_sse2_sin" } } */
+/* { dg-final { scan-assembler "__libm_sse2_sinf" } } */
